requirements.dist.txt
django>=2.0.0,<3.0.0
djangorestframework>=3.8.0,<4.0.0